private integration tests
http://github.com/astriaorg/sequencer-relayer/pull/33

In #24 `types::Header::to_tendermint_header` was made public so it can be run as as an integration test. This is not ideal because we don't want it to be in our public interfaces.

This PR makes it crate-private `pub(crate)` again and introduces `src/tests.rs` for tests that are like integration tests but which are testing private objects.

Because `sequencer-relayer/src/tests.rs` and and `sequencer-relayer/tests/` now share common initialization routines the entire crate was made into a workspace with:
+ `sequencer-relayer` containing the same logic as before, and
+ `sequencer-relayer-test` containing the initialization logic that was formerly in `sequencer-releayer/tests/helper.rs`.

Note further that this PR introduces a `.cargo/config.toml` that defines an env var `CARGO_WORKSPACE_DIR` as per https://github.com/rust-lang/cargo/issues/3946 so that build scripts (for protobuf generation) and `include_str!` (in the tests setup) can find their respective sources.