implement startupProbes and wait on bridge
http://github.com/astriaorg/sequencer-relayer/pull/34

This PR should finally fix flaky tests.

It introduces `startupProbe`s for celestia bridge and metro (just hitting `/header/1` and `/cosmos/base/tendermint/v1beta1/blocks/latest`) so that kubernetes does not consider the deployment `"Available"` until after these have returned a success. In the absence of proper `/readyz` endpoints this is the best next thing.

Because there is a tiny delay (about 20ms on my machine) between the deployment coming online and this also introduces a loop inside the tests that attemps to hit `/bridge/header/1` until it returns something in the range `200-399` every seconds.