use tracing for telemetry (#34)
This commit replaces log and flexilogger by tracing and tracing-subscriber.
The default log level is `info`, which can be overriden by using the `RUST_LOG`
environment variable with directives as described in the tracing subscriber docs:
https://docs.rs/tracing-subscriber/0.3.17/tracing_subscriber/filter/struct.EnvFilter.html

Because env var directives are strictly more powerful than setting verbosity levels
on the command line the `log-level` flag is removed.

If astria-conductor is connected to a tty it will emit tracing events using
human-readable format. If it is not connected to a tty it will emit JSON.

By default logging is disabled in integration tests. It can be enabled by
setting the environment variable `TEST_LOG=1`.

Note that the service no longer writes explicitly writes logs to a file.
Because astria-conductor is intended to be run as a service by a service manager
or orchestration platform, it is assumed that persisting logs is handled
externally.