turn per-crate into workspace dependencies (#38)
Crate dependencies are made into workspace dependencies if used in more than one crate.

`rand` is kept at 0.7 for now, because `ed25519-dalek:1.0.1` still requires it. It's bumped to 0.8 in `ed25519-dalek:rc.2`, but a) is that not a full release, and b) that breaks `libp2p-noise`'s `snow` dependency.

Once `libp2p` is updated to 0.52 we should be able to update rand 0.8.