container: speedup docker builds by only building binary dependencies (#172)
This commit adds the `--bin` arguments to the `cargo chef prepare` and
`cargo chef cook` steps in order to speed up compilation by only
compiling workspace members that are needed for building the target
binary.

Note that at the time of this commit only speedups for
`astria-sequencer` are noticable, because `astria-conductor` and
`astria-sequencer-relayer` transitively pull in most of the other
workspace members. This is tracked in issue #191.

Because cargo-chef version 0.1.61 contains a bug when preparing the
initial recipe, this commit uses the argument `--bin
crates/$TARGETBINARY` in the prep phase, but `--bin $TARGETBINARY` in
the cook phase.