composer: initial implementation (#192)
This commit adds a minimal composer service with the binary
`astria-composer`. It is currently restricted to the following
functionality:

1. it subscribes to new pending transactions of an EVM rollup using the
`eth_subscribe` JSONRPC with parameters `["newPendingTransactions",
true]`;
2. it wraps the evm transaction in a sequencer transaction, and submits
it to the sequencer using a new randomly generated account;
3. it exposes a `/readyz` HTTP endpoint that returns a status code of
200 and `{ "status": "code" }` if it was able to execute a
`node_version` JSONRPC against the EVM rollup and a `abci_info` against
sequencer/cometbft.

This functionality was tested using a blackbox test with mocked services
in place of the EVM rollup and the sequencer.

Composer is configured via environment variables only. It does not
provide command line arguments nor a config file.

Build and run astria composer with the following commands (assuming the
sequencer/cometbft service and execution/EVM are listening on their
standard ports):

```sh
$ cargo build --release -p astria-composer
$ ASTRIA_COMPOSER_LOG="astria_composer=info" \
   ASTRIA_COMPOSER_API_LISTEN_ADDR="0.0.0.0:8080" \
   ASTRIA_COMPOSER_SEQUENCER_URL="http://127.0.0.1:26657" \
   ASTRIA_COMPOSER_CHAIN_ID="912559" \
   ASTRIA_COMPOSER_EXECUTION_URL="ws://127.0.0.1:8545" \
   ./target/release/astria-composer
```