feat(sequencer): implement generation of action tree commitment (#209)
first part of #153 (implementing solution 4. in that issue):
> 4. do the sorting on the block proposer side, create a commitment to
the batch of txs with data for a specific rollup, and put that as a
special "transaction" type inside a block. then, the flow will be:
> 
> a. conductor fetches batch of data for its rollup + the special
"commitment transaction" + merkle proof of commitment tx inside
data_hash
> b. re-calculates and verifies the commitment using its batch of data
> c. verifies that the commitment was included inside the block's
data_hash
this seems like a good solution, but we'll need to update the
application validator logic to verify that the commitment txs are valid
commitments before voting yes for that block.

- implement generation of "action tree" and put the merkle root of this
tree as the first "transaction" in a block when proposing
- verify this when receiving a proposal
- remove `IndexedTransaction` in favour of just `Vec<u8>` as tx indices
are no longer relevant for tx verification

next steps are #208 and then implementing the actual verification of the
batch of rollup txs on the conductor side by verifying that it's
included in the commitment transaction.