fix(sequencer): fix `deliver_tx` to accept the action tree root (#291)
## Summary

my previous PR #209 actually broke sequencer+cometbft since `deliver_tx`
is called on every tx, after `prepare_proposal`. since it was trying to
deserialize every tx into `Signed`, it was failing on the action tree
commitment (as it's just 32 bytes at the start of the block)

## Changes

- update the sequencer `App` to handle the first tx in a block properly,
right now it just checks that it's 32 bytes. this is fine because the
content of these 32 bytes is checked in `process_proposal`.

## Testing

ran it with cometbft and now works

## Related Issues

related to #153