feat(sequencer): implement generating and verifying chain ID commitment (#338)
## Summary
proposers now make a commitment to the `sequence::Action` `chain_id`s
inside the block. this will allow for conductors to verify that the
sequencer-relayer has not omitted any namespaces in the
`SequencerBlockData` they write. (punishment if they do is TBD)

## Background
- required so that we can verify that the sequencer-relayer did not
totally omit writing a rollup's data for a block

## Changes
- proposers now make a commitment to the `sequence::Action` `chain_id`s
inside the block inside `prepare_proposal` (placed as the second tx in a
block, directly after the current commitment to `sequence::Action`s)
- other validators verify this commitment in `process_proposal`

## Testing
unit tests

## Breaking Changelist
- proposed blocks must now contain the chain ID commitment as the second
tx in the block

## Related Issues

part 1 of #308 (follow up will implement this verification in conductor)