feat(conductor): implement verification of chain IDs commitment (#365)
## Summary
previous PR implemented generating the chain IDs commitment in
`prepare_proposal` and verifying it in `process_proposal`. this PR
finishes the job by ensuring the chain IDs commitment is verified in
conductor.

## Background
required so that we can ensure that the sequencer-relayer actually wrote
data for every rollup ID in a block, previously it would have been able
to omit data for a whole rollup without detection. we still need to
implement slashing (eventually) for this - what happens right now is
that the conductor will fail to verify a block if the commitment does
not match up, even if it was able to obtain the rollup data for its own
chain ID.

## Changes
-  ensuring the chain IDs commitment is verified in conductor
- update the sequencer app and
`SequencerBlockData::from_tendermint_block` to handle a block with 2
commitments at the start of the txs correctly; this should have been
done in my previous PR, apologies
- update ChainId to be max 32 bytes

## Testing
unit tests

## Related Issues

closes #308