sequencer-validation: implement std Error (#430)
## Summary
Enable the `std` feature on `ct-merkle`, fix the std `Error`
implementation for `VerificationFailure`.

## Background
A wellformed std error should have a correct implementation of its
`source` trait method. This was not possible because it was thought that
`std::error::Error` is not implemented for the inner error. It actually
is, but hidden behind a feature flag.

## Changes
- Enable `std` feature on `ct-merkle` dependency
- Fix `std::error::Error for VerificationFailure` implementation.

## Testing
Not necessaty

## Breaking Changelist
- Nothing externally observable.

## Related Issues
closes https://github.com/astriaorg/astria/issues/374