feat(sequencer): implement support for arbitrary assets (#568)
## Summary
This PR implements support for arbitrary assets for fee payments and
balance/transfers inside the sequencer. Currently, only one asset exists
(the native asset, RIA), but after IBC is implemented we can potentially
have many arbitrary assets on the sequencer chain.

## Background
We need to support multiple assets for IBC, and also support fee
payments in different token types (probably ibc-TIA)

## Changes
- implement a `Denom` type, which represents the base (smallest)
denomination of an arbitrary asset
- implement an `Id` type, which is a hash of the denomination "trace"
(see
https://github.com/cosmos/ibc-go/blob/main/docs/architecture/adr-001-coin-source-tracing.md
for details) in this PR, since there is no IBC, the `Id` is just the
hash of the base denomination.
- update `UnsignedTransaction` to have an optional `fee_asset` field; if
this is unset, the native asset is used for fee payment
- update `Transfer` to have an optional `asset` field; if this is unset,
the native asset is the asset transferred
- update the sequencer app to have a `NATIVE_ASSET` global
- update the genesis state to have a `native_asset_base_denomination`
field. the native asset should be set in the genesis file, and stored in
(non-consensus) storage for retrieval upon startups. then the
`NATIVE_ASSET` global is set upon startup by reading from storage.

## Testing
unit tests, ran it locally

## Breaking Changelist
- `UnsignedTransaction` and `Transfer` now have an extra (optional)
field.

## Related Issues

closes #559