refactor: enforce sequencer blob invariants, simplify conductor (#576)
## Summary
Enforce invariants when constructing sequencer blobs.

## Background
The sequencer blob type `SequencerNamespaceData` that was written to and
read from celestia was treated as a dumb container, delegating
consistency checks to downstream users (for example, to the
`block_verifier` module in `astria-conductor`). In the spirit of making
invalid states unrepresentable, this patch introdues a
`RawSequencerNamespaceData` as the dumb (de)serialization container,
from which a verified `SequencerNamespaceData` can be built. In
particular, the following invariants are upheld at construction,
rejecting a conversion otherwise (`SMD` stands for
`SequencerNamespaceData` and `SMD.field` refers to the `field` of the
same):

1. `SMD.header.data_hash` must be set
2. `SMD.block_hash` must match the result of `SMD.header.hash()`
3. `SMD.action_tree_root` and `SMD.action_tree_root_inclusion_proof`
must verify against `SMD.data_hash`
4. The Merkle tree hash built from `SMD.rollup_chain_ids` must match
`SMD.chain_ids_commitment`
5. The `SMD.chain_ids_commitment` and `SMD.chain_ids_commitment_proof`
must verifiy against `SMD.data_hash`

With these changes conductor's `BlockVerifier` can be simplified
significantly and is now only responsible for checking the validity of a
given sequencer blob against the remote sequencer.

## Changes
- Introduce `RawSequencerNamespaceData` type
- Make all `SequencerNamespaceData` fields private, add read-only
getters
- Change `SequencerNamespaceData` serialization to be in terms of the
`Raw*` type
- Enforce sequencer blob invariants by introducing
`SequencerNamespaceData::try_from_raw` as the only fallible constructor
- Add a `SMD::data_hash` getter for convenience which panics if the
invariants are violated
- Streamline `BlockVerifier` by removing all checks that now happen
during `SequencerNamespaceData` construction

## Testing
How are these changes tested?

## Related Issues
This is done as part of #356 

Blocked on #554 being merged