feat: redefine sequencer blocks, celestia blobs as protobuf (#395)
## Summary
This patch defines protobuf versions of various types that are currently
only defined in Rust and exchange as json-serialized text.

## Background
`astria-proto` is intended to be the single source of truth for the wire
format of data that is exchanged between services. At the moment
`SequencerBlockData`, `SequencerNamespaceData`, `RollupNamespaceData`
and `InclusionProof` are only defined in Rust. Some older versions of
these exist as protobuf but were never used in the public interfaces.

## Changes
- Replace `SequencerBlockData` by a protobuf `SequencerBlock` (rewriting
this protobuf type is technically a breaking change, but it was never
used, so in practice it is not)
- Replace `SequencerNamespaceData` and `RollupNamespaceData` by protobuf
`CelestiaSequencerBlob` and `CelestiaRollupBlob`.
- Introduce other protobuf types like `RollupTransactions`, `Proof` to
decode/encode all required objects
- Write idiomatic rust types for all generated protobuf representations
to uphold invariants

## Testing
- all tests have been updated and still pass
- testing this is in a devnet to check that the data still flows is open

## Related Issues

https://github.com/astriaorg/astria/pull/12 originally defined sequencer
blocks and transactions as protobuf, but these were never actually
implemented.

https://github.com/astriaorg/astria/issues/86 is the reason this effort
was started.

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>